doc: update limitations for match with compare result item
In switch mode with enabled `repr_matching_en` devarg flag, the match
with compare result item is not supported for `ingress` rules. This is
due to an implicit REPRESENTED_PORT being added to the matcher, which
conflicts with the single item limitation for match with compare result
item.

Update the documentation to reflect this limitation in the MLX5 driver
guide.

Fixes: cb25df7ce9d6 ("net/mlx5: support comparison matching")
Cc: stable@dpdk.org

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>
Acked-by: Bing Zhao <bingz@nvidia.com>