app/testpmd: use RSS conf from software when configuring DCB
In the testpmd command, we have to stop the port firstly before configuring
the DCB. However, some PMDs may execute a hardware reset during the port
stop, such as ixgbe. Some kind of reset operations of PMD could clear the
configurations of RSS in the hardware register. This would cause the loss
of RSS configurations that were set during the testpmd initialization. As
a result, I find that I cannot enable RSS and DCB at the same time in the
testpmd command when using Intel 82599 NIC.

The patch uses rss conf from software instead of reading from the hardware
register when configuring DCB.

Signed-off-by: Min Zhou <zhoumin@loongson.cn>
Acked-by: Stephen Hemminger <stephen@networkplumber.org>